+++ libs/cx/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.49.0-rc.43](https://github.com/SAP/fundamental-ngx/compare/v0.49.0-rc.42...v0.49.0-rc.43) (2024-03-12)
+
+**Note:** Version bump only for package @fundamental-ngx/cx
+
+
+
+
+
+# [0.49.0-rc.42](https://github.com/SAP/fundamental-ngx/compare/v0.49.0-rc.41...v0.49.0-rc.42) (2024-03-12)
+
+**Note:** Version bump only for package @fundamental-ngx/cx
+
+
+
+
+
 # [0.49.0-rc.41](https://github.com/SAP/fundamental-ngx/compare/v0.49.0-rc.40...v0.49.0-rc.41) (2024-03-11)
 
 **Note:** Version bump only for package @fundamental-ngx/cx